refactor: modify directory structure
moved BridgeGame to model directory
